pyproject.toml
[project]
name = "tgcalendar"
version = "0.1.0"
description = "Telegram bot with Mini App calendar for personal and family tasks"
authors = [
    { name = "TGCalendar Team", email = "you@example.com" }
]
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.4.1",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-dotenv>=1.0.1",
    "pydantic[email]",
    "pydantic-settings>=2.4.0",
    "SQLAlchemy>=2.0.30",
    "alembic>=1.13.1",
    "itsdangerous>=2.2.0",
    "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0"
]

[tool.uv]
dev-dependencies = []

[tool.setuptools.packages.find]
include = ["app*", "bot*"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
